requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.23
pdfkit==0.6.1
psycopg2-binary==2.8.6
PyMySQL==1.0.2
Werkzeug==1.0.1
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pandas==1.2.4
pyodbc==4.0.28






